+++ src/test/java/racingcar/strategy/FixtureMoveStrategy.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.strategy;
 
 import racingcar.strategy.MoveStrategy;
